requirements.txt
fastapi==0.102.0
uvicorn[standard]==0.23.2
tensorflow==2.12.0
opencv-python-headless==4.12.0.88
pillow==10.0.1
numpy==1.25.2
requests==2.32.0
